refactor(mcp): extract agent HITL decision handling from RunMCPToolReactor
